/summary commit during 19/12/2022
