+++ src/library.html
@@ -16,6 +16,7 @@
   <include src="./partials/header-library.html"></include>
   <include src="./partials/gallery.html"></include>
   <include src="./partials/pagination.html"></include>
+  <script src="https://kit.fontawesome.com/633f3e6c9a.js" crossorigin="anonymous"></script>
   <script type="module" src="./index.js"></script>
   <script type="module" src="/src/js/switchLanguage.js"></script>
 </body>
